refactor(server-remote-cloudflare): inline storage.ts into yjs-room.ts
