+++ packages/server-remote-cloudflare/src/yjs-room.ts
@@ -11,7 +11,7 @@ import {
 	type UpdateLog,
 } from '@epicenter/sync-core';
 import * as Y from 'yjs';
-import { createDoSqliteUpdateLog } from './storage';
+
 
 type WsAttachment = {
 	controlledClientIds: number[];
@@ -214,4 +214,63 @@ export class YjsRoom extends DurableObject {
 	override async webSocketError(ws: WebSocket, _error: unknown): Promise<void> {
 		await this.webSocketClose(ws, 1011, 'WebSocket error', false);
 	}
+}
+
+/**
+ * Create an UpdateLog backed by Durable Object SQLite.
+ *
+ * Uses the DO's built-in SQLite database for persistent Y.Doc update storage.
+ * SQLite in Durable Objects is GA with 10GB per DO.
+ */
+function createDoSqliteUpdateLog(
+	storage: DurableObjectStorage,
+): UpdateLog {
+	let initialized = false;
+
+	function ensureTable() {
+		if (initialized) return;
+		storage.sql.exec(`
+			CREATE TABLE IF NOT EXISTS updates (
+				id INTEGER PRIMARY KEY AUTOINCREMENT,
+				doc_id TEXT NOT NULL,
+				data BLOB NOT NULL,
+				created_at INTEGER DEFAULT (unixepoch())
+			)
+		`);
+		initialized = true;
+	}
+
+	return {
+		append(docId, update) {
+			ensureTable();
+			storage.sql.exec(
+				'INSERT INTO updates (doc_id, data) VALUES (?, ?)',
+				docId,
+				update,
+			);
+		},
+
+		readAll(docId) {
+			ensureTable();
+			const cursor = storage.sql.exec(
+				'SELECT data FROM updates WHERE doc_id = ? ORDER BY id',
+				docId,
+			);
+			return [...cursor].map(
+				(row) => new Uint8Array(row.data as ArrayBuffer),
+			);
+		},
+
+		replaceAll(docId, mergedUpdate) {
+			ensureTable();
+			storage.transactionSync(() => {
+				storage.sql.exec('DELETE FROM updates WHERE doc_id = ?', docId);
+				storage.sql.exec(
+					'INSERT INTO updates (doc_id, data) VALUES (?, ?)',
+					docId,
+					mergedUpdate,
+				);
+			});
+		},
+	};
 }
